requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
Pillow==8.2.0
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
PyJWT~=2.0.1